feat: 딜러 클래스 분리 및 히트 가능 여부 판단하는 기능 구현
