+++ src/main/kotlin/blackjack/domain/Participant.kt
@@ -1,0 +1,29 @@
+package blackjack.domain
+
+abstract class Participant(val name: String) {
+
+    private val _cards = mutableListOf<Card>()
+    val cards: List<Card>
+        get() = _cards.toList()
+
+    fun receive(card: Card) {
+        _cards.add(card)
+    }
+
+    fun getScore(): Int {
+        var score = _cards.sumOf { it.value }
+
+        val aceCount = _cards.count { it.isAce() }
+        repeat(aceCount) { score = updateScore(score) }
+        return score
+    }
+
+    fun isBusted(): Boolean = getScore() > TARGET_SCORE
+
+    private fun updateScore(score: Int): Int = if (score > TARGET_SCORE) score - GAP_ACE else score
+
+    companion object {
+        private const val TARGET_SCORE = 21
+        private val GAP_ACE: Int = CardNumber.ACE.value - 1
+    }
+}
